bot_telegram/requirements.txt
python-telegram-bot==20.6
boto3
python-dotenv
requests

